remove finalize request, and move reward to finalize request update status event
